Fix propertiesbased & separate tests run in CI
Issue #207
With the ServiceLoader, the tests needs to be run with the concerned
module as dependency. This implied adding a dependency to the pcaptest
module to load the FactoryBinderProvider when running the tests.
Now the tests are not run inside each module but are run by activating a
specific profile on the pcaptest module.

I also reflected the package changes in the packet-factory.properties
file